requirements.txt
coverage==7.6.12
Django==5.1.6
django-debug-toolbar==5.0.1
django-filter==25.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pytest-django==4.10.0
pytest-pythonpath==0.7.4
python-dotenv==1.0.1
requests==2.32.3
